requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.7
django-bootstrap5==23.3
django-forms-bootstrap==3.1.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
idna==3.4
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
psycopg2-binary==2.9.3
